+++ rust/integration-tests/src/diff_test_helper/token_v2_processor.rs
@@ -1,4 +1,4 @@
-use crate::models::queryable_models::TokenActivityV2;
+use crate::queryable_models::token_v2_models::TokenActivityV2;
 use anyhow::Result;
 use diesel::{
     pg::PgConnection,
