--- rust/integration-tests/src/models/mod.rs
@@ -1,1 +1,0 @@
-pub mod queryable_models;
