+++ rust/integration-tests/src/queryable_models/fungible_asset_models.rs
@@ -7,31 +7,10 @@ use bigdecimal::BigDecimal;
 use diesel::{Identifiable, Insertable, Queryable};
 use field_count::FieldCount;
 use processor::schema::{
-    coin_supply, current_fungible_asset_balances_legacy, events, fungible_asset_activities,
-    fungible_asset_balances, fungible_asset_metadata, token_activities_v2,
+    coin_supply, current_fungible_asset_balances_legacy, fungible_asset_activities,
+    fungible_asset_balances, fungible_asset_metadata,
 };
 use serde::{Deserialize, Serialize};
-/**
-* Event model
-* this is created // b/c there is inserated_at field which isn't defined in the Event struct, we can't just load the events directly without specifying the fields.
-        // TODO: make this more generic to load all fields, then we should be able to run tests for all processor in one test case.
-
-*/
-#[derive(Clone, Debug, Deserialize, FieldCount, Identifiable, Insertable, Serialize, Queryable)]
-#[diesel(primary_key(transaction_version, event_index))]
-#[diesel(table_name = events)]
-pub struct Event {
-    pub sequence_number: i64,
-    pub creation_number: i64,
-    pub account_address: String,
-    pub transaction_version: i64,
-    pub transaction_block_height: i64,
-    pub type_: String,
-    pub data: serde_json::Value,
-    pub inserted_at: chrono::NaiveDateTime,
-    pub event_index: i64,
-    pub indexed_type: String,
-}
 
 #[derive(Clone, Debug, Deserialize, FieldCount, Identifiable, Insertable, Serialize, Queryable)]
 #[diesel(primary_key(transaction_version, event_index))]
@@ -54,28 +33,6 @@ pub struct FungibleAssetActivity {
     pub transaction_timestamp: chrono::NaiveDateTime,
     pub inserted_at: chrono::NaiveDateTime,
     pub storage_refund_amount: BigDecimal,
-}
-
-#[derive(Clone, Debug, Deserialize, FieldCount, Identifiable, Insertable, Serialize, Queryable)]
-#[diesel(primary_key(transaction_version, event_index))]
-#[diesel(table_name = token_activities_v2)]
-pub struct TokenActivityV2 {
-    pub transaction_version: i64,
-    pub event_index: i64,
-    pub event_account_address: String,
-    pub token_data_id: String,
-    pub property_version_v1: BigDecimal,
-    pub type_: String,
-    pub from_address: Option<String>,
-    pub to_address: Option<String>,
-    pub token_amount: BigDecimal,
-    pub before_value: Option<String>,
-    pub after_value: Option<String>,
-    pub entry_function_id_str: Option<String>,
-    pub token_standard: String,
-    pub is_fungible_v2: Option<bool>,
-    pub transaction_timestamp: chrono::NaiveDateTime,
-    pub inserted_at: chrono::NaiveDateTime,
 }
 
 #[derive(Clone, Debug, Deserialize, FieldCount, Identifiable, Insertable, Serialize, Queryable)]
