+++ rust/integration-tests/src/queryable_models/mod.rs
@@ -1,0 +1,3 @@
+pub mod events_models;
+pub mod fungible_asset_models;
+pub mod token_v2_models;
